Moved seleced and expanded to redux

+++ 802_11_cdb_app/src/Users.js
@@ -1,29 +1,30 @@
-import PropTypes from 'prop-types';
-import React, {useState, useEffect, useRef} from 'react';
-import { connect } from 'react-redux';
-import AppTable from './AppTable';
-import AppModal from './AppModal';
-import {sortClick, filterValidate} from './filter';
-import {ActionButton} from './Icons';
-import {setUsersFilters, setUsersSort, getUsers, updateUser, addUser, deleteUsers, uploadUsers} from './actions/users';
+import PropTypes from 'prop-types'
+import React, {useState, useEffect, useRef} from 'react'
+import { connect } from 'react-redux'
+import AppTable from './AppTable'
+import AppModal from './AppModal'
+import {sortClick, filterValidate} from './filter'
+import {ActionButton} from './Icons'
+import ConfirmModal from './ConfirmModal'
+import {setUsersFilters, setUsersSort, setUsersSelected, getUsers, updateUser, addUser, deleteUsers, uploadUsers} from './actions/users'
 
 
 function AddUserModal(props) {
-	const defaultUserData = {SAPIN: '', Name: '', Email: '', Access: 3};
-	const [userData, setUserData] = useState(defaultUserData);
+	const defaultUserData = {SAPIN: '', Name: '', Email: '', Access: 3}
+	const [userData, setUserData] = useState(defaultUserData)
 
 	function onOpen() {
 		setUserData(defaultUserData)
 	}
 
 	function change(e) {
-		setUserData({...userData, [e.target.name]: e.target.value});
+		setUserData({...userData, [e.target.name]: e.target.value})
 	}
 
 	function submit(e) {
-		console.log(userData);
-		props.dispatch(addUser(userData));
-		props.close();
+		console.log(userData)
+		props.dispatch(addUser(userData))
+		props.close()
 	}
 
 	return (
@@ -54,7 +55,7 @@ AddUserModal.propTypes = {
 }
 
 function UploadUsersModal(props) {
-	const usersFileInputRef = useRef();
+	const usersFileInputRef = useRef()
 
 	function submit() {
 		props.dispatch(uploadUsers(usersFileInputRef.current.files[0])).then(props.close)
@@ -107,88 +108,90 @@ function Users(props) {
 			width: 100,
 			cellRenderer: renderAccess,
 			isLast: true}
-	];
+	]
 	const primaryDataKey = columns[0].dataKey
 
-	const [showAddUserModal, setShowAddUserModal] = useState(false);
-	const [showUploadUsersModal, setShowUploadUsersModal] = useState(false);
-	const [selected, setSelected] = useState([])
+	const [showAddUserModal, setShowAddUserModal] = useState(false)
+	const [showUploadUsersModal, setShowUploadUsersModal] = useState(false)
 
 	const [tableSize, setTableSize] = useState({
 		height: 400,
 		width: 300,
-	});
+	})
 
 	function updateTableSize() {
 		const maxWidth = columns.reduce((acc, col) => acc + col.width, 0)
-		const headerEl = document.getElementsByTagName('header')[0];
+		const headerEl = document.getElementsByTagName('header')[0]
 
-		const height = window.innerHeight - headerEl.offsetHeight - 5;
-		const width = window.innerWidth - 1;
+		const height = window.innerHeight - headerEl.offsetHeight - 5
+		const width = window.innerWidth - 1
 
 		if (height !== tableSize.height || width !== tableSize.width) {
-			setTableSize({height, width: Math.min(width, maxWidth)});
+			setTableSize({height, width: Math.min(width, maxWidth)})
 		}
 	}
 	
 	useEffect(() => {
-		updateTableSize();
-		window.addEventListener("resize", updateTableSize);
+		updateTableSize()
+		window.addEventListener("resize", updateTableSize)
 		return () => {
-			window.removeEventListener("resize", updateTableSize);
+			window.removeEventListener("resize", updateTableSize)
 		}
 	}, [])
 
 	useEffect(() => {
 		if (Object.keys(props.filters).length === 0) {
-			var filters = {};
+			var filters = {}
 			for (let col of columns) {
 				if (col.filterable) {
 					filters[col.dataKey] = filterValidate(col.dataKey, '')
 				}
 			}
-			props.dispatch(setUsersFilters(filters));
+			props.dispatch(setUsersFilters(filters))
 		}
-		if (!props.usersDataValid) {
+		if (!props.usersValid) {
 			props.dispatch(getUsers())
 		}
 	}, [])
 
-	function handleRemoveSelected() {
-		const {usersData, usersDataMap} = props;
-		var ids = [];
-		for (var i = 0; i < usersDataMap.length; i++) { // only select checked items that are visible
-			let id = usersData[usersDataMap[i]][primaryDataKey]
+	async function handleRemoveSelected() {
+		const {users, usersMap, selected} = props
+		var ids = []
+		for (var i = 0; i < usersMap.length; i++) { // only select checked items that are visible
+			let id = users[usersMap[i]][primaryDataKey]
 			if (selected.includes(id)) {
 				ids.push(id)
 			}
 		}
 		if (ids.length) {
-			props.dispatch(deleteUsers(ids))
+			const ok = await ConfirmModal.show('Are you sure you want to delete ' + ids.join(', ') + '?')
+			if (ok) {
+				await props.dispatch(deleteUsers(ids))
+			}
 		}
 	}
 
 	function refresh() {
-		props.dispatch(getUsers());
+		props.dispatch(getUsers())
 	}
 
 	function updateUserField(rowIndex, dataKey, fieldData) {
-		const usersDataIndex = props.usersDataMap[rowIndex];
-		const u = props.usersData[usersDataIndex];
-		props.dispatch(updateUser({
-			UserID: u.UserID,
-			[dataKey]: fieldData
-		}));
+		const index = props.usersMap[rowIndex]
+		const u = props.users[index]
+		if (dataKey === 'SAPIN') {
+			fieldData = parseInt(fieldData, 10)
+		}
+		props.dispatch(updateUser(u.SAPIN, {[dataKey]: fieldData}))
 	}
 
 	function updateUserFieldIfChanged(rowIndex, dataKey, fieldData) {
-		const usersDataIndex = props.usersDataMap[rowIndex];
-		const u = props.usersData[usersDataIndex];
+		const index = props.usersMap[rowIndex]
+		const u = props.users[index]
+		if (dataKey === 'SAPIN') {
+			fieldData = parseInt(fieldData, 10)
+		}
 		if (u[dataKey] !== fieldData) {
-			props.dispatch(updateUser({
-				UserID: u.UserID,
-				[dataKey]: fieldData
-			}));
+			props.dispatch(updateUser(u.SAPIN, {[dataKey]: fieldData}))
 		}
 	}
 
@@ -221,13 +224,13 @@ function Users(props) {
 	}
 
   	function setSort(dataKey, event) {
-		const {sortBy, sortDirection} = sortClick(event, dataKey, props.sortBy, props.sortDirection);
-		props.dispatch(setUsersSort(sortBy, sortDirection));
+		const {sortBy, sortDirection} = sortClick(event, dataKey, props.sortBy, props.sortDirection)
+		props.dispatch(setUsersSort(sortBy, sortDirection))
 	}
 
 	function setFilter(dataKey, value) {
 		var filter = filterValidate(dataKey, value)
-		props.dispatch(setUsersFilters({[dataKey]: filter}));
+		props.dispatch(setUsersFilters({[dataKey]: filter}))
 	}
 
 	return (
@@ -236,29 +239,26 @@ function Users(props) {
 				<span><label>Users</label></span>
 				<span>
 					<ActionButton name='add' title='Add User' onClick={() => setShowAddUserModal(true)} />
-					<ActionButton name='delete' title='Remove Selected' onClick={handleRemoveSelected} />
+					<ActionButton name='delete' title='Remove Selected' disabled={props.selected.length === 0} onClick={handleRemoveSelected} />
 					<ActionButton name='upload' title='Upload Users' onClick={() => setShowUploadUsersModal(true)} />
 					<ActionButton name='refresh' title='Refresh' onClick={refresh} />
 				</span>
 			</div>
 			<AppTable
-				hasRowSelector={true}
-				hasRowExpander={false}
 				columns={columns}
 				rowHeight={22}
 				height={tableSize.height}
 				width={tableSize.width}
 				loading={props.getUsers}
-				filters={props.filters}
 				sortBy={props.sortBy}
 				sortDirection={props.sortDirection}
 				setSort={setSort}
+				filters={props.filters}
 				setFilter={setFilter}
-				//showSelected={() => setShowSelected(true)}
-				setSelected={(ids) => setSelected(ids)}
-				selected={selected}
-				data={props.usersData}
-				dataMap={props.usersDataMap}
+				selected={props.selected}
+				setSelected={(ids) => props.dispatch(setUsersSelected(ids))}
+				data={props.users}
+				dataMap={props.usersMap}
 				primaryDataKey={primaryDataKey}
 			/>
 
@@ -267,6 +267,7 @@ function Users(props) {
 				close={() => setShowAddUserModal(false)}
 				dispatch={props.dispatch}
 			/>
+			
 			<UploadUsersModal
 				isOpen={showUploadUsersModal}
 				close={() => setShowUploadUsersModal(false)}
@@ -277,18 +278,20 @@ function Users(props) {
 }
 
 function mapStateToProps(state) {
-	const s = state.users;
+	const s = state.users
 	return {
 		filters: s.filters,
 		sortBy: s.sortBy,
 		sortDirection: s.sortDirection,
-		usersDataValid: s.usersDataValid,
-		usersData: s.usersData,
-		usersDataMap: s.usersDataMap,
+		selected: s.selected,
+		usersValid: s.usersValid,
+		users: s.users,
+		usersMap: s.usersMap,
 		getUsers: s.getUsers,
 		updateUser: s.updateUser,
 		addUser: s.addUser,
 		deleteUsers: s.deleteUsers,
 	}
 }
-export default connect(mapStateToProps)(Users);
+
+export default connect(mapStateToProps)(Users)
